+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixes a OAuth login bug for the queues page when using queues.spellbot.io.
+
 ### Added
 
 - Adds the Champ de Bataille server.